Rename openssl_interface namespace to bridge
